first attempt at creating a user input component
in this file so far made the inputs fields apear according to the type of input.
started to make it so that the data labels are linked to the inputs from start.
prevented the page from reloading when submmiting the form.
turned the string input of select and group to arrays to be called later on data